UPDATE SAN PHAM
xong giao dien. chuẩn bi cập nhat csdl

+++ class/upload.php
@@ -6,21 +6,24 @@
  * Time: 2:06 PM
  * Usage: upload hinh anh len website (thay đổi logo và thêm hinh ảnh sản phẩm)
  */
-
+include_once('./config.php');
 class upload{
     private $name='';
-    private $path='./products/';
+    private $path='';
 
    public function upload($input_name)
     {
         $this->name=$input_name;
+        $this->path=ROOT.'/products/';
     }
 
-    public function add_products()
+    public function add_products($old_file=NULL)
     {
-
+        if(!is_null($old_file)) {
+            $this->del_products($old_file);
+        }
         $uploads=$this->path.$this->name;
-        move_uploaded_file($_FILES['hinh']['tmp_name'],$uploads);
+        move_uploaded_file($_FILES['Hinh']['tmp_name'],$uploads);
     }
     public function change_logo($output_name)
     {
@@ -28,10 +31,14 @@ class upload{
         unlink($uploads);
         move_uploaded_file($_FILES[$this->name]['tmp_name'],$uploads);
     }
-    public function del_products()
+    public function del_products($old_file=NULL)
     {
-        $uploads=$this->path.$this->name;
-        unlink($uploads);
+        if(!is_null($old_file)) {
+            $uploads=$this->path.$old_file;
+            unlink($uploads);
+        }
+        else{$uploads=$this->path.$this->name;
+        unlink($uploads);}
     }
 }
